`CRAB_CAST_CENSUS`: a third of `cube`'s casts pay the take-from-hand round trip for nothing
PERF candidate (Q), filed off this run's CoW census and censused the same day,
which is what this branch's instruments are for. `cast_spell_with_convoke` is
the program's largest CoW unsharer — 22,182 of 103,388 `make_mut_slow` calls,
0.85 % of `cube` — and the reason is structural: `actions.rs` takes the card
**out of hand** at the top and then runs its rejection checks, forty-nine of
which put it back, each having already unshared `PlayerData`, the hand `Vec`
and the card.

    CRAB_CAST_CENSUS=1 bot_ladder --a gang --b gang --games 6 --threads 1 --decks <pool>

              calls   reached the take   put the card back
    cube      7,660        100.0 %        2,550  = 33.3 %
    sealed    8,574        100.0 %        1,700  = 19.8 %
    fixed     3,566        100.0 %           72  =  2.0 %

Counted, not inferred: the rollback bump sits at each `hand.push(card)` rather
than being derived from an `Err` return, so the retry path (the suspend-and-
replay at the end of the body, which is where 1,748 of the 9,408 calls come
from) is counted with the rejections — it pays the same round trip.

⚠ **Every call reaches the take: the removal is guarded by nothing.** At the
CoW family's own prices the ceiling is ~0.30 % of `cube`, the largest single
row left on the queue — and `fixed`'s 2.0 % says it will read as nothing on
the pool `--bench` measures, so rank it on `cube`. PERF's (Q) carries the
plan: about twenty of the forty-nine rollbacks sit in the first 38 % of the
body and every one of them reads `card.definition` or the board rather than the
owned card, so they can be hoisted above the removal — one group per commit,
suite between, census re-run after each.

The instrument matches `attack_census` / `menu_census` (an `OnceLock` gate, an
`AtomicU64` array, a `snapshot()` printed by `bot_ladder`) and is documented in
PERF's "How to measure". Also fixes two `unnecessary_cast` clippy warnings that
arrived with `143980a7`.

Suite 19,603 / 0 / 6 with `CRAB_ANSWER_LOG=strict`, clippy 0 over the
workspace, `cargo check --profile release-fast` clean.

+++ crabomination/src/server/bot.rs
@@ -16356,8 +16356,8 @@ fn x_candidates(state: &GameState, seat: usize, action: &GameAction, x: u32) -> 
     }
     let cap = (state.effective_life(seat) - 1).max(0) as u32;
     let mut v: Vec<u32> = vec![0];
-    v.extend(state.battlefield.iter().map(|c| c.printed_cmc() as u32));
-    v.extend(state.players[seat].graveyard.iter().map(|c| c.printed_cmc() as u32));
+    v.extend(state.battlefield.iter().map(|c| c.printed_cmc()));
+    v.extend(state.players[seat].graveyard.iter().map(|c| c.printed_cmc()));
     v.retain(|&m| m <= cap && m != x);
     v.sort_unstable();
     v.dedup();
